Create new wallpapersplitter class
The dialog and old wallpaper splitter will be merged into this
