+++ main.cpp
@@ -1,6 +1,6 @@
 #include <QApplication>
 #include "offsetdialog.h"
-#include "wallpapersplitter.h"
+#include "wallpapersplitter_old.h"
 
 int main(int argc, char *argv[]) {
     // TODO add support for non-gui version
