+++ wallpapersplitter.ui
@@ -1,60 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>WallpaperSplitter</class>
- <widget class="QMainWindow" name="WallpaperSplitter">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>277</width>
-    <height>299</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Wallpaper Splitter</string>
-  </property>
-  <widget class="QWidget" name="centralwidget">
-   <layout class="QVBoxLayout" name="verticalLayout">
-    <item>
-     <widget class="QPushButton" name="imageButton">
-      <property name="text">
-       <string>Select a wallpaper image</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="offsetButton">
-      <property name="enabled">
-       <bool>false</bool>
-      </property>
-      <property name="text">
-       <string>Adjust offset</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="splitButton">
-      <property name="enabled">
-       <bool>false</bool>
-      </property>
-      <property name="text">
-       <string>Split the image</string>
-      </property>
-     </widget>
-    </item>
-    <item>
-     <widget class="QPushButton" name="applyButton">
-      <property name="enabled">
-       <bool>false</bool>
-      </property>
-      <property name="text">
-       <string>Apply</string>
-      </property>
-     </widget>
-    </item>
-   </layout>
-  </widget>
- </widget>
- <resources/>
- <connections/>
+    <author/>
+    <comment/>
+    <exportmacro/>
+    <class>WallpaperSplitter</class>
+    <widget class="QWidget" name="WallpaperSplitter">
+        <property name="geometry">
+            <rect>
+                <x>0</x>
+                <y>0</y>
+                <width>400</width>
+                <height>300</height>
+            </rect>
+        </property>
+        <property name="windowTitle">
+            <string>WallpaperSplitter</string>
+        </property>
+    </widget>
+    <pixmapfunction/>
+    <connections/>
 </ui>
